% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.Rrs.from.buoy.R
\name{compute.Rrs.from.buoy}
\alias{compute.Rrs.from.buoy}
\title{Compute remote sensing reflectance from uncorrected optical data from
the IML (VIKING) buoys using Bélanger et al (2016) algorithm}
\usage{
compute.Rrs.from.buoy(waves, Ed0, LuZ, DateTime, lon, lat)
}
\arguments{
\item{waves}{Wavelenghts of Ed0 and LuZ}

\item{Ed0}{Spectral irradiance above water}

\item{LuZ}{Spectral upwelling radiance measured by the sensor at 0.86m
below sea surface.}

\item{DateTime}{A date and time string as POSIXct object}

\item{lon}{Longitude in degrees (negative for west),}

\item{lat}{Latitude in degrees}
}
\value{
Returns a data.frame with sun zenith angle (sunzen),
remote sensing refectances(Rrs),
water leaving radiance (Lw),
normalized water leaving radiance (nLw),
 and reflectance (rhown)
}
\description{
Apply shadow correction, extrapol from 0.86m to surface and compute Rrs from raw optical data of IML Buoy
}
\details{
The data processing includes mainly two corrections.
First the upwelling radiance is corrected for shadow effect using
an empirical algorithm developped at the IML4 site during summer 2015.
The algorithm is detailed in Bélanger et al
(manuscript in preparation, June 2016).
}
\examples{

#read an optical data file from IML Buoy
setwd(path.package("Rimlbuoy"))
raw = read.OPTICD("extdata/IML-4_OPTICD_20160501_000320.DAT")

# select an indice. here 65 is 16:30 GMT or 12:30 t local time
ix = 65
rrs = compute.Rrs.from.buoy(raw$waves,raw$Ed0p[ix,], raw$Lu0.86m[ix,], raw$DateTime[ix], raw$lon[ix], raw$lat[ix])

# plot the Rrs spectrum
plot(raw$waves, rrs$Rrs, type="l", main=raw$DateTime[ix], ylim=c(0,max(rrs$Rrs, na.rm=TRUE)),sub=paste("Latitude:", signif(raw$lat[ix],6), "   Longitude:", signif(raw$lon[ix],6)), xlab="Wavelength")


}

