% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_str_coord.R
\name{convert_str_coord}
\alias{convert_str_coord}
\title{Convert coordinates degrees minutes secondes to decimal}
\usage{
convert_str_coord(c)
}
\arguments{
\item{c}{is character coordinates following "48°40.21'N" format.}
}
\value{
decimal coordinates
}
\description{
Convert coordinates degrees minutes secondes to decimal
}
\examples{
convert_str_coord("48°40.21'N")
convert_str_coord("068°34.75'W")

}
\author{
Thomas Jaegler, modified by Simon Bélanger
}

