% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.bandshift.coefficient.R
\name{get.bandshift.coefficient}
\alias{get.bandshift.coefficient}
\title{Get bandshift coefficient to remote sensing reflectance data from
the IML (VIKING) buoys using Zibordi et al (2009) algorithm}
\usage{
get.bandshift.coefficient(Rrs0, waves0, band.width0, waves, band.width)
}
\arguments{
\item{Rrs0}{: reference remote sensing reflectance that need to be coorected for bandshift (numeric vector).}

\item{waves0}{: reference wavelength (numeric vector).}

\item{band.width0}{: bandwidth of the reference wavelength (numeric vector).}

\item{waves}{: target wavelength for the bandshift correction (numeric vector).}

\item{band.width}{: bandwidth of a target wavelength (numeric vector).}
}
\value{
Return list of data with the bandshift coefficient The length of
the list is the length of the reference wavelength. Each element have the length of
the sequence center on the targeted wavelength (bandwith dependant).
}
\description{
Get bandshift coefficient to remote sensing reflectance data from
the IML (VIKING) buoys using Zibordi et al (2009) algorithm
}
\details{
The correction use the coefficient of the linear regression between
in-situ remote sensing ratio and in-situ IOP measurement of bbp, Ag, A_{nap},
A_{ph} from the 2015 mission on the IML4 buoy. The correction protocol is describe
in Zibordi et al. (2009). The coefficient of the linear regression are used to estimate empirical
modelisation of the IOP to calculate total absorption and total backscattering. Then the ratio of
IOP and reference remote sensing reflectance is used as Eq.7 of Zibordi et al. (2009) to apply the
bandshift corrections.

  \deqn{Rrs(\lambda)= Rrs(\lambda 0)*\frac{b_{b}(\lambda)}{a(\lambda)+b_{b}(\lambda)}*\frac{a(\lambda 0)+b_{b}(\lambda 0)}{b_{b}(\lambda 0)}}{%
  Rrs(\lambda)= Rrs(\lambda 0)
  *b_{b}(\lambda)/(a(\lambda)+b_{b}(\lambda))
  *(a(\lambda 0)+b_{b}(\lambda 0))/b_{b}(\lambda 0)}


 References :


 Zibordi, G., et al., Validation of satellite ocean color primary products at optically complex coastal sites: Northern
 Adriatic Sea, Northern Baltic Proper and Gulf of Finland, Remote Sensing of Environment (2009).
}
\examples{

#read an optical data file from IML Buoy
setwd(path.package("Rimlbuoy"))
raw = read.OPTICD("data/IML-4_OPTICD_20160501_000320.DAT")

#
waves0=c(412,443,490,510,555,670, 683)
band.width0 = rep(10,7)

waves = c(560,665,681)
band.width = rep(10,3)
ix = 65
rrs = compute.Rrs.from.buoy(raw$waves,raw$Ed0p[ix,], raw$Lu0.86m[ix,], raw$DateTime[ix], raw$lon[ix], raw$lat[ix])

 bdsfh.coef=get.bandshift.coefficient(rrs$Rrs,waves0,band.width0,waves,band.width)




}
\author{
Thomas Jaegler
}

