% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/METOCE.merge.R
\name{METOCE.merge}
\alias{METOCE.merge}
\title{Merge daily METOCE files to create time series}
\usage{
METOCE.merge(path, StartDate, EndDate)
}
\arguments{
\item{path}{is the path where METOCE files are located}

\item{StartDate}{is the starting date of the time series
(format is yyyymmdd)}

\item{EndDate}{is the ending date of the time series
(format is yyyymmdd)}
}
\value{
Returns a list similar to the \code{\link{read.METOCE}}
}
\description{
Merge daily METOCE files to create time series
}
\author{
Simon Bélanger
}

