% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPTICD.merge.R
\name{OPTICD.merge}
\alias{OPTICD.merge}
\title{Merge daily OPTICD files to create multi days time series}
\usage{
OPTICD.merge(path, StartDate, EndDate)
}
\arguments{
\item{path}{is the path where OPTICD files are located}

\item{StartDate}{is the starting date of the time series
(format is yyyymmdd)}

\item{EndDate}{is the ending date of the time series
(format is yyyymmdd)}
}
\value{
Returns a list similar to the \code{\link{read.OPTICD}}
}
\description{
The OPTICD file contains OCR data as well as
fluorescence (FDOM and FCHL), GPS, and
tilt and heading information.
}
\author{
Simon Bélanger
}

