% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.METOCE.R
\name{read.METOCE}
\alias{read.METOCE}
\title{Reads METOCE file coming from IML Buoy}
\usage{
read.METOCE(filen)
}
\arguments{
\item{fname}{Filename}
}
\value{
It returns a list of fields containing all the data
stored in the METOCE file.
}
\description{
The METOCE file contains 20 parameters: meteorological data
(wind speed, raffale, wind direction,
air temperature, pressure, PAR),
oceanographic data (water temperature, conductivity, salinity),
fluorescence (FDOM and FCHL), GPS, and tilt and heading information.
}
\author{
Simon Bélanger
}

