% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.OCR3.R
\name{read.OCR3}
\alias{read.OCR3}
\title{Reads optics data coming from IML Buoy and returns a list of field containing all the data stored in the OCR3 file. (see example)}
\usage{
read.OCR3(filen)
}
\arguments{
\item{fname}{Filename}
}
\value{
list of data
Headers :
DateTime   : Time Format DD-MMM-YYYY HH:MM:SS.ss
DOY        : Counter
lat        : Latitude (North +ve)
lon        : Longitude (East +ve)
waves      : Wavelength standard
}
\description{
Reads optics data coming from IML Buoy and returns a list of field containing all the data stored in the OCR3 file. (see example)
}
\examples{
# read an optical data file from IML Buoy dat file
setwd(path.package("Rimlbuoy"))
raw = read.OCR3("extdata/IML-4_OCR3_20160501_000320.DAT")
str(raw)




}

