% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.OCR_MED.R
\name{read.OCR_MED}
\alias{read.OCR_MED}
\title{Reads optics data coming from IML Buoy and returns a list of field containing all the data stored in the OPTIC file. (see example)}
\usage{
read.OCR_MED(filen)
}
\arguments{
\item{fname}{Filename}
}
\value{
list of data
Headers :
DateTime   : Time Format DD-MMM-YYYY HH:MM:SS.ss
DOY        : Counter
lat        : Latitude (North +ve)
lon        : Longitude (East +ve)
waves      : Wavelength standard
Ed0p       : Irradiance for each wavelength
Ed0p.sd    : Standard deviation of irradiance for each wavelength
Lu0.86m    : Radiance for each wavelength
Lu0.86m.sd : Standard deviation of radiance for each wavelength
PAR        : Photosynthetic Active Radiation
PAR.sd     : Standard deviation of Photosynthetic Active Radiation
FCHL       : Fluorescence of Chlorophyl
FCHL.sd    : Standard deviation of Fluorescence of Chlorophyl
FDOM       : Fluorescence of CDOM (Coloured Dissolved Organic Matter)
FDOM.sd    : Standard deviation of Fluorescence of CDOM (Coloured Dissolved Organic Matter)
pitch      : Pitch Angle
pitch.sd   : Standard deviation of Pitch Angle
roll       : Roll Angle
roll.sd    : Standard deviation of Roll Angle
tilt       : Tilt Angle
tilt.sd    : Standard deviation of Tilt Angle
heading    : Heading
}
\description{
Reads optics data coming from IML Buoy and returns a list of field containing all the data stored in the OPTIC file. (see example)
}
\examples{
# read an optical data file from IML Buoy dat file
setwd(path.package("Rimlbuoy"))
raw = read.OCR_MED("extdata/IML-4_OCR_MED_20160501_000320.DAT")
str(raw)




}

