% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.OPTICD.R
\name{read.OPTICD}
\alias{read.OPTICD}
\title{Reads OPTICD file coming from IML Buoy}
\usage{
read.OPTICD(filen)
}
\arguments{
\item{fname}{Filename}
}
\value{
It returns a list of fields containing all the data
stored in the OPTICD file. (see example)

Headers :
DateTime   : Time Format DD-MMM-YYYY HH:MM:SS.ss
DOY        : Counter
lat        : Latitude (North +ve)
lon        : Longitude (East +ve)
waves      : Wavelength standard
Ed0p       : Irradiance for each wavelength
Ed0p.sd    : Standard deviation of irradiance for each wavelength
Lu0.86m    : Radiance for each wavelength
Lu0.86m.sd : Standard deviation of radiance for each wavelength
PAR        : Photosynthetic Active Radiation
PAR.sd     : Standard deviation of Photosynthetic Active Radiation
FCHL       : Fluorescence of Chlorophyl
FCHL.sd    : Standard deviation of Fluorescence of Chlorophyl
FDOM       : Fluorescence of CDOM (Coloured Dissolved Organic Matter)
FDOM.sd    : Standard deviation of Fluorescence of CDOM (Coloured Dissolved Organic Matter)
pitch      : Pitch Angle
pitch.sd   : Standard deviation of Pitch Angle
roll       : Roll Angle
roll.sd    : Standard deviation of Roll Angle
tilt       : Tilt Angle
tilt.sd    : Standard deviation of Tilt Angle
heading    : Heading
}
\description{
The OPTICD file contains OCR data as well as
fluorescence (FDOM and FCHL), GPS, and tilt and heading information.
}
\examples{
# read an optical data file from IML Buoy dat file
setwd(path.package("Rimlbuoy"))
raw = read.OPTICD("extdata/IML-4_OPTICD_20160501_00032")
str(raw)


}
\author{
Simon Bélanger
}

