% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.PAR_DIGI.R
\name{read.PAR_DIGI}
\alias{read.PAR_DIGI}
\title{Reads PAR_DIGI file coming from IML Buoy}
\usage{
read.PAR_DIGI(filen)
}
\arguments{
\item{fname}{Filename}
}
\value{
It returns a list of fields containing all the data
stored in the PAR_DIGI file. (see example)

Headers :
DateTime   : Time Format DD-MMM-YYYY HH:MM:SS.ss
DOY        : Counter
lat        : Latitude (North +ve)
lon        : Longitude (East +ve)
pitch      : Pitch Angle
pitch.sd   : Standard deviation of Pitch Angle
roll       : Roll Angle
roll.sd    : Standard deviation of Roll Angle
tilt       : Tilt Angle
tilt.sd    : Standard deviation of Tilt Angle
}
\description{
The PAR_DIGI file contains PAR and tilt data from the Mini buoy.
}
\examples{
# read an optical data file from IML Buoy dat file
setwd(path.package("Rimlbuoy"))
raw = read.PAR_DIGI("extdata/IML-")
str(raw)


}
\author{
Thomas Jaegler
}

