% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.SD.R
\name{read.SD}
\alias{read.SD}
\title{Reads meteo data coming from IML Buoy and returns a list of field containing all the data stored in the SD file. (see example)}
\usage{
read.SD(filen)
}
\arguments{
\item{fname}{Filename}
}
\value{
file
Headers :
DateTime   : Time Format DD-MMM-YYYY HH:MM:SS.ss
lat        : Latitude (North +ve)
lon        : Longitude (East +ve)
WSP        : Wind speed
RW         : Draft speed
DirW       : Direction wind
TAIR       : Temperatur air
HUM        : Humidity
PRES       : Pressure
TWAT       : Water temperature
COND       : Conductivity
SAL        : Salinity
DENS       : Density
WETS       : Wetstar
WETS_sd    : Wetstar standard deviation
CDOM       : CdOM fluo
CDOM_sd    : CDOM fluo standard deviation
PAR        : Photosyntetical active radiation
PAR_sd     : Photosyntetical active radiation standard deviation
}
\description{
Reads meteo data coming from IML Buoy and returns a list of field containing all the data stored in the SD file. (see example)
}
\examples{
# read an SD data file from IML Buoy dat file
setwd(path.package("Rimlbuoy"))
raw = read.SD("data/IML-4_SD_20160501_000300.DAT")
str(raw)



}

