% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.to.MERMAID.R
\name{write.to.MERMAID}
\alias{write.to.MERMAID}
\title{Generate an ASCII file for the MERMAID data base containing
dark corrected remote sensing reflectance and water leaving radiance from the
IML buoy.}
\usage{
write.to.MERMAID(OPTICD_fname, outputdir, Tilt_threshold = 10,
  USE.MEDIAN = FALSE, PI, PI_Email, land_dist_IS, PQC, MQC,
  USE.BANDSHIFT.COEF = TRUE, band.width, waves2, band.width2)
}
\arguments{
\item{OPTICD_fname}{Name of the OPTICD data file IML buoy containing
mean and standard deviation for each OCR channels.}

\item{outputdir}{Path for the ouptut directory where the
results will be saved in ASCII format.}

\item{Tilt_threshold}{is the threshold for maximum tilt angle tolerance.
Default value is 10.}

\item{USE.MEDIAN}{is a logical parameter indicating whether the Rrs is
cumputed using median values stored OCR_MED files. If TRUE, the file name
of the output will begin by rrs_med_.
Default is FALSE.}

\item{PI}{: pi full name}

\item{PI_Email}{: pi email adress}

\item{land_dist_IS}{: In-situ measurement distance to the land}

\item{PQC}{: Processing quality control from MERMAID (see MERMAID data format report from ACRI-ST)}

\item{MQC}{: Measurement Quality Control from MERMAID (see MERMAID data format report from ACRI-ST)}

\item{USE.BANDSHIFT.COEF}{is a logical parameter indicating whether the bandshift coefficient must
be used if TRUE the following parameter are needed (band.width,waves2,band.width2).
Default is TRUE.}

\item{band.width}{: bandwidth of the reference wavelength (numeric vector).}

\item{waves2}{: target wavelength for the bandshift correction (numeric vector).}

\item{band.width2}{: bandwidth of a target wavelength (numeric vector).}
}
\value{
It writes an csv file names rrs_DATE.csv (or rrs_med_DATE.csv) and
returns a data frame in the command line.
}
\description{
This routine takes two files coming from the buoys as input, then apply
a number of quality control check and compute Rrs following the method detailed
in Belanger et al.
}
\details{
This function reads the data file provided as input,
apply quality control checks and then compute Rrs using \code{\link{compute.Rrs.from.buoy}}.
Next ìf USE.BANDSHIFT.COEF is true, the bandshift correction is apply to nlw and rho_w. Then the
flags for coefficient of variation (<20%) and tilt (<10°) are identified and excluded. Then the final
data frame is formated with MERMAID header.
}
\examples{
#Define file names
setwd(path.package("Rimlbuoy"))
pathwd="extdata/"
fnames = list.files(pathwd,full.names = T)
fn=fnames[grep("OPTICD",fnames)]

#Define threshold for tilt flag
Tilt_threshold = 10
band.width=rep(10,7)
waves2= c(560,665,681)
band.width2 = rep(10,3)
PI="Simon_Belanger"
PI_Email ="sbe@arctus.ca"
land_dist_IS = 20
PQC = "P000010000"
MQC="M110111001101020200"


#Define output dir
output_dir_data="extdata/"

#Run write.to.MERMAID
res = write.to.MERMAID(OPTICD_fname=fn,outputdir=output_dir_data,Tilt_threshold=Tilt_threshold)
res.med = write.to.MERMAID(OPTICD_fname=fn,outputdir=output_dir_data,Tilt_threshold=Tilt_threshold,USE.MEDIAN=T)
str(res)

}
\author{
Thomas Jaegler modified by Simon Belanger
}

